I can write the create portfolio name to the database and I can write the relationship of the stock from the ui button.
I need the current portfolio instance passed to that button event method
and we will have it. 

Also we need to display the stockks and the portfolios and have the
current instance of the portfolio in the text box (this will allow all
the methods that need it a way to grab the current insdstance from main.